feat: move modal components to ui/modals components
