+++ project/components/ui/buttons/create-list-button.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { CreateListModal } from "@/components/modals/create-list-modal";
+import { CreateListModal } from "@/components/ui/modals/create-list-modal";
 import { ProjectSchema } from "@/types/projects";
 import { PlusSquare } from "lucide-react";
 import { Fragment, useState } from "react";
