+++ project/components/ui/modals/create-project-modal.tsx
@@ -1,10 +1,10 @@
 import { createProject } from "@/lib/actions/projects";
 import { redirect, RedirectType } from "next/navigation";
 import { useState } from "react";
-import Button from "../ui/buttons/button";
-import Input from "../ui/input-fields/input";
-import Textarea from "../ui/input-fields/textarea";
-import { showErrorToast, showSuccessToast } from "../ui/toast";
+import Button from "../buttons/button";
+import Input from "../input-fields/input";
+import Textarea from "../input-fields/textarea";
+import { showErrorToast, showSuccessToast } from "../toast";
 import Modal from "./modal";
 
 interface CreateProjectModalProps {
