+++ project/components/ui/modals/edit-project-modal.tsx
@@ -35,10 +35,10 @@ import { updateProject } from "@/lib/actions/projects";
 import { ProjectSchema } from "@/types/projects";
 import { redirect, RedirectType } from "next/navigation";
 import { useEffect, useState } from "react";
-import Button from "../ui/buttons/button";
-import Input from "../ui/input-fields/input";
-import Textarea from "../ui/input-fields/textarea";
-import { showErrorToast, showSuccessToast } from "../ui/toast";
+import Button from "../buttons/button";
+import Input from "../input-fields/input";
+import Textarea from "../input-fields/textarea";
+import { showErrorToast, showSuccessToast } from "../toast";
 import Modal from "./modal";
 
 interface EditProjectModalProps {
